+++ projects/core/src/cms/config/default-cms-config.ts
@@ -1,9 +1,6 @@
 import { CmsConfig } from './cms-config';
 
 export const defaultCmsModuleConfig: CmsConfig = {
-  cmsComponents: {
-    CMSTabParagraphComponent: { selector: 'cx-paragraph' },
-  },
   backend: {
     occ: {
       endpoints: {
