+++ projects/storefrontlib/src/cms-components/user/login-form/login-form.module.ts
@@ -10,8 +10,8 @@ import {
   UrlModule,
   UserModule,
 } from '@spartacus/core';
+import { CmsModule } from '../../../cms-structure/cms.module';
 import { BootstrapModule } from '../../../lib/bootstrap.module';
-import { CmsModule } from '../../../lib/cms/cms.module';
 import { LoginFormComponent } from './login-form.component';
 
 @NgModule({
