+++ projects/storefrontlib/src/cms-structure/guards/index.ts
@@ -1,0 +1,1 @@
+export * from './cms-page.guard';
