+++ projects/storefrontlib/src/cms-structure/index.ts
@@ -1,3 +1,5 @@
+export * from './cms.module';
+export * from './guards/index';
 export * from './outlet/index';
 export * from './page/index';
 export * from './pwa/index';
