+++ projects/storefrontlib/src/cms-structure/routing/cms-route/add-cms-route.ts
@@ -1,6 +1,6 @@
 import { Injector } from '@angular/core';
 import { Route, Router } from '@angular/router';
-import { CmsPageGuard } from '../../../lib/cms/guards/cms-page.guard';
+import { CmsPageGuard } from '../../guards/cms-page.guard';
 import { PageLayoutComponent } from '../../page/index';
 
 const cmsRoute: Route = {
