+++ projects/storefrontlib/src/layout/main/storefront.component.html
@@ -1,3 +1,6 @@
+<ng-template cxOutletRef="SearchBox">
+  test
+</ng-template>
 <header
   [class.is-expanded]="isExpanded$ | async"
   (keydown.escape)="collapseMenu()"
