--- projects/storefrontlib/src/lib/cms/guards/index.ts
@@ -1,5 +1,0 @@
-import { CmsPageGuard } from './cms-page.guard';
-
-export const guards: any[] = [CmsPageGuard];
-
-export * from './cms-page.guard';
