--- projects/storefrontlib/src/lib/cms/index.ts
@@ -1,4 +1,0 @@
-export * from './cms.module';
-export * from './guards/cms-page.guard';
-export * from './services/cms-mapping.service';
-export * from './services/cms-routes.service';
