+++ projects/storefrontlib/src/lib/storefront.module.ts
@@ -11,10 +11,10 @@ import {
   StateModule,
 } from '@spartacus/core';
 import { CmsLibModule } from '../cms-components/index';
+import { CmsModule } from '../cms-structure/cms.module';
 import { CmsRouteModule } from '../cms-structure/routing/cms-route/cms-route.module';
 import { SuffixRoutesModule } from '../cms-structure/routing/suffix-routes/suffix-routes.module';
 import { MultiStepCheckoutModule } from './checkout/index';
-import { CmsModule } from './cms/index';
 import { provideConfigFromMetaTags } from './provide-config-from-meta-tags';
 import { StorefrontModuleConfig } from './storefront-config';
 import { UiModule } from './ui/index';
