+++ projects/storefrontlib/src/public_api.ts
@@ -7,7 +7,6 @@ export * from './cms-components/myaccount/index';
 export * from './cms-structure/index';
 export * from './layout/index';
 export * from './lib/checkout/index';
-export * from './lib/cms/index';
 export * from './lib/storefront-config';
 export * from './lib/storefront.module';
 export * from './lib/ui/index';
